v1.5.41 - Speed up wanted mode by reducing processing time
Previous process built the album list, and then created a second list from the original in lowercase format to improve matching by removing case sensitivity. One negative effect was the process could take a very long time being that it is single threaded process and a lot of data to be processed at once.

This changes the caching process to essentially double cache the data for each item. One in normal case and a second one that is converted to lowercase. This results in an almost near instant speed increase because the reduction in the amount of data to be converted.